+++ src/Event/Results/ResultCard.tsx
@@ -1,12 +1,12 @@
 import { StyleSheet, View, Text } from "react-native"
 import { useTheme } from "@react-navigation/native"
 
-import { ImageType } from "../types";
-import PlaceholderImage from "../Shared/PlaceholderImage";
-import { Standing, User } from "./types"; 
-import { getNumberOrdinal } from "../helper";
+import { ImageType } from "../../types";
+import PlaceholderImage from "../../Shared/PlaceholderImage";
+import { Standing, User } from "../types";
+import { getNumberOrdinal } from "../../helper";
 
-function getImages(participants: {user: User}[]): string[] {
+function getImages(participants: { user: User }[]): string[] {
     const images = participants.map((participant) => {
         const user = participant.user;
         const image = user.images[0];
@@ -18,7 +18,7 @@ function getImages(participants: {user: User}[]): string[] {
 
 
 
-const ResultCard = ({playerData, index}: {playerData: Standing, index: number}) => {
+const ResultCard = ({ playerData, index }: { playerData: Standing, index: number }) => {
 
     const { colors } = useTheme();
     const placement = playerData.placement;
@@ -37,8 +37,8 @@ const ResultCard = ({playerData, index}: {playerData: Standing, index: number}) 
         const player = playerData.player;
         const user = player.user;
         const image = user.images[0];
-        const profileImage = image ? image.url : '';        
-    
+        const profileImage = image ? image.url : '';
+
         return (
             <View style={containerStyle}>
                 <View style={styles.imageContainer}>
@@ -46,11 +46,11 @@ const ResultCard = ({playerData, index}: {playerData: Standing, index: number}) 
                 </View>
                 <View style={styles.detailsContainer}>
                     <View style={styles.playerTitle}>
-                        <Text style={{...styles.playerTag, color: colors.text}}>{player.gamerTag}</Text>
-                        { player.prefix && <Text style={{...styles.playerSponsor, color: colors.secondaryText}}>{player.prefix}</Text> }
-                        { user.genderPronoun && <Text style={{...styles.playerPronoun, color: colors.secondaryText}}>{user.genderPronoun}</Text>}
+                        <Text style={{ ...styles.playerTag, color: colors.text }}>{player.gamerTag}</Text>
+                        {player.prefix && <Text style={{ ...styles.playerSponsor, color: colors.secondaryText }}>{player.prefix}</Text>}
+                        {user.genderPronoun && <Text style={{ ...styles.playerPronoun, color: colors.secondaryText }}>{user.genderPronoun}</Text>}
                     </View>
-                    <Text style={{...styles.playerPlacement, color: colors.text}}>{placementString}</Text>
+                    <Text style={{ ...styles.playerPlacement, color: colors.text }}>{placementString}</Text>
                 </View>
             </View>
         )
@@ -60,17 +60,17 @@ const ResultCard = ({playerData, index}: {playerData: Standing, index: number}) 
     const entrant = playerData.entrant;
     const participants = entrant.participants;
     const images = getImages(participants);
-    
+
     return (
         <View style={containerStyle}>
             <View style={styles.imageContainer}>
-                <PlaceholderImage imageSrc={images[0]} placeholder="player" style={styles.image}/>
+                <PlaceholderImage imageSrc={images[0]} placeholder="player" style={styles.image} />
             </View>
             <View style={styles.detailsContainer}>
                 <View style={styles.playerTitle}>
-                    <Text style={{...styles.playerTag, color: colors.text}}>{entrant.name}</Text>
+                    <Text style={{ ...styles.playerTag, color: colors.text }}>{entrant.name}</Text>
                 </View>
-                <Text style={{...styles.playerPlacement, color: colors.text}}>{placementString}</Text>
+                <Text style={{ ...styles.playerPlacement, color: colors.text }}>{placementString}</Text>
             </View>
         </View>
     )
