+++ src/Event/Results/ResultsPage.tsx
@@ -3,14 +3,14 @@ import { ActivityIndicator, FlatList, StyleSheet, Text, ToastAndroid, View } fro
 import { useTheme } from "@react-navigation/native";
 
 import ResultCard from "./ResultCard";
-import { Entrant, EventAPIQuery } from "../types";
-import { EventStandingsQuery, queryAPI } from "../api";
-import SearchBar from "../Shared/SearchBar";
+import { Entrant, EventAPIQuery } from "../../types";
+import { EventStandingsQuery, queryAPI } from "../../api";
+import SearchBar from "../../Shared/SearchBar";
 
 const PER_PAGE = 24    // Show 24 players per page (instead of the default 25).
 
-const ResultsPage = ({navigation, route}) => {
-    
+const ResultsPage = ({ navigation, route }) => {
+
     // UI state 
     const [refresh, setRefresh] = useState(false);
     const [updating, setUpdating] = useState(false);
@@ -25,7 +25,7 @@ const ResultsPage = ({navigation, route}) => {
     const eventId = route.params.id;
     const singles = route.params.singles;
     const { colors } = useTheme();
-    
+
     async function addPlacements() {
         setUpdating(true);
 
@@ -74,34 +74,34 @@ const ResultsPage = ({navigation, route}) => {
     }, [page])
 
     useEffect(() => {
-        if (finished && page > 1){
+        if (finished && page > 1) {
             console.info("No more players");
             ToastAndroid.show("No more players", ToastAndroid.SHORT);
         }
     }, [finished])
 
     return (
-        <View style={{flex: 1}}>
+        <View style={{ flex: 1 }}>
             <FlatList
                 style={styles.container}
-                ListHeaderComponent={ <SearchBar setFilter={setFilter} filterAction={filterEntrants} searchTitle="Search" style={{marginTop: 20}}/> }
+                ListHeaderComponent={<SearchBar setFilter={setFilter} filterAction={filterEntrants} searchTitle="Search" style={{ marginTop: 20 }} />}
                 data={standings}
-                renderItem={({index, item}) => <ResultCard playerData={item} index={index}/>}
-                
-                contentContainerStyle={{flexGrow: 1}}
-                ListEmptyComponent={ 
-                    <View style={styles.centerText}>
-                        <Text style={{color: colors.text}}>No entrants were found</Text>
-                    </View> }
-                
-                onEndReached={() => setPage(page+1)}
-                onEndReachedThreshold={0.1}
-                />
+                renderItem={({ index, item }) => <ResultCard playerData={item} index={index} />}
 
-            { updating && 
-            <View style={styles.loadingCircle}>
-                <ActivityIndicator size='large' color={colors.primary} />
-            </View> }
+                contentContainerStyle={{ flexGrow: 1 }}
+                ListEmptyComponent={
+                    <View style={styles.centerText}>
+                        <Text style={{ color: colors.text }}>No entrants were found</Text>
+                    </View>}
+
+                onEndReached={() => setPage(page + 1)}
+                onEndReachedThreshold={0.1}
+            />
+
+            {updating &&
+                <View style={styles.loadingCircle}>
+                    <ActivityIndicator size='large' color={colors.primary} />
+                </View>}
         </View>
     )
 }
